src/cf-api-to-s3/requirements.txt
functions-framework==3.7.0
pandas==2.2.1
requests==2.32.3
pyarrow==15.0.2
s3fs==2024.2.0
boto3==1.34.22
google-cloud-secret-manager==2.26.0
python-dotenv==1.0.1
